screen: Refactor level
